# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BioGene Unified Bioinformatics Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /protein/search:
    get:
      operationId: searchProteins
      tags: [protein]
      summary: Search proteins by name, gene, or organism via RCSB PDB
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProteinSearchResponse"
...
  /nlp/papers/search:
    get:
      operationId: searchPubmedPapers
      tags: [nlp]
      summary: Search PubMed for open-access papers
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Papers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PubmedSearchResponse"
...
  /genome/search:
    get:
      operationId: searchGenomicRegion
      tags: [genome-browser]
      summary: Search for genes or features in the genome
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: species
          in: query
          required: false
          schema:
            type: string
            default: human
      responses:
        "200":
          description: Genomic features
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenomicSearchResponse"
...
  /drugs/search:
    get:
      operationId: searchCompounds
      tags: [drugs]
      summary: Search drug compounds from ChEMBL
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Compound search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompoundSearchResponse"
...
  /drugs/targets/search:
    get:
      operationId: searchTargets
      tags: [drugs]
      summary: Search drug targets from ChEMBL
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Target search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TargetSearchResponse"
...
  /transcriptomics/expression/search:
    get:
      operationId: searchGeneExpression
      tags: [transcriptomics]
      summary: Search gene expression data from GTEx/Ensembl
      parameters:
        - name: gene
          in: query
          required: true
          schema:
            type: string
        - name: tissue
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Expression data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpressionSearchResponse"
...
  /search:
    get:
      operationId: globalSearch
      tags: [search]
      summary: BM25 natural-language search across all datasets with field:value filters
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: collections
          in: query
          required: false
          schema:
            type: string
          description: Comma-separated collection names to scope the search
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Ranked search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
...
  /search/facets:
    get:
      operationId: getSearchFacets
      tags: [search]
      summary: Facet counts for a field, optionally scoped by collection or query
      parameters:
        - name: field
          in: query
          required: true
          schema:
            type: string
        - name: collections
          in: query
          required: false
          schema:
            type: string
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 25
      responses:
        "200":
          description: Facet counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchFacetsResponse"
...
components:
  schemas:
...
    ProteinEntry:
      type: object
      properties:
        pdbId:
          type: string
        title:
          type: string
        organism:
          type: string
        resolution:
          type: number
        method:
          type: string
        depositionDate:
          type: string
        releaseDate:
          type: string
        citationCount:
          type: integer
        citations:
          type: array
          items:
            $ref: "#/components/schemas/Citation"
        chains:
          type: integer
        atoms:
          type: integer
        ligands:
          type: array
          items:
            type: string
        uniprotId:
          type: string
        gene:
          type: string
        pdbUrl:
          type: string
      required: [pdbId, title]
    ProteinSearchResponse:
      type: object
      properties:
        proteins:
          type: array
          items:
            $ref: "#/components/schemas/ProteinEntry"
        total:
          type: integer
      required: [proteins, total]
...
    PubmedPaper:
      type: object
      properties:
        pmid:
          type: string
        title:
          type: string
        abstract:
          type: string
        authors:
          type: array
          items:
            type: string
        journal:
          type: string
        year:
          type: integer
        doi:
          type: string
        keywords:
          type: array
          items:
            type: string
        citationCount:
          type: integer
        pubmedUrl:
          type: string
        doiUrl:
          type: string
      required: [pmid, title]
    PubmedSearchResponse:
      type: object
      properties:
        papers:
          type: array
          items:
            $ref: "#/components/schemas/PubmedPaper"
        total:
          type: integer
        query:
          type: string
      required: [papers, total, query]
...
    GenomicFeature:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        chromosome:
          type: string
        start:
          type: integer
        end:
          type: integer
        strand:
          type: integer
        biotype:
          type: string
        description:
          type: string
        organism:
          type: string
        ncbiGeneUrl:
          type: string
        ensemblUrl:
          type: string
      required: [id, name, type, chromosome, start, end]
    GenomicSearchResponse:
      type: object
      properties:
        features:
          type: array
          items:
            $ref: "#/components/schemas/GenomicFeature"
        total:
          type: integer
        species:
          type: string
      required: [features, total]
...
    Compound:
      type: object
      properties:
        chemblId:
          type: string
        name:
          type: string
        synonyms:
          type: array
          items:
            type: string
        smiles:
          type: string
        inchi:
          type: string
        molecularFormula:
          type: string
        molecularWeight:
          type: number
        alogp:
          type: number
        hbondDonors:
          type: integer
        hbondAcceptors:
          type: integer
        rotatableBonds:
          type: integer
        aromaticRings:
          type: integer
        qedScore:
          type: number
        ro5Violations:
          type: integer
        psa:
          type: number
        maxPhase:
          type: integer
        indication:
          type: string
        atcClass:
          type: string
        atcClasses:
          type: array
          items:
            type: string
        firstApproval:
          type: integer
        moleculeType:
          type: string
        chemblUrl:
          type: string
        pubchemUrl:
          type: string
        indications:
          type: array
          items:
            type: object
            properties:
              term:
                type: string
              maxPhase:
                type: number
            required: [term]
        totalIndications:
          type: integer
      required: [chemblId, name]
    CompoundSearchResponse:
      type: object
      properties:
        compounds:
          type: array
          items:
            $ref: "#/components/schemas/Compound"
        total:
          type: integer
        query:
          type: string
      required: [compounds, total, query]
...
    DrugTarget:
      type: object
      properties:
        targetChemblId:
          type: string
        targetName:
          type: string
        targetType:
          type: string
        organism:
          type: string
        geneNames:
          type: array
          items:
            type: string
        uniprotId:
          type: string
        chemblUrl:
          type: string
        uniprotUrl:
          type: string
      required: [targetChemblId, targetName]
    TargetSearchResponse:
      type: object
      properties:
        targets:
          type: array
          items:
            $ref: "#/components/schemas/DrugTarget"
        total:
          type: integer
      required: [targets, total]
...
    GeneExpressionEntry:
      type: object
      properties:
        geneId:
          type: string
        geneName:
          type: string
        tissue:
          type: string
        tissueName:
          type: string
        ontologyId:
          type: string
        tpm:
          type: number
        median:
          type: number
        unit:
          type: string
        gtexUrl:
          type: string
      required: [geneId, geneName, tissue, tpm]
    ExpressionSearchResponse:
      type: object
      properties:
        gene:
          type: string
        geneId:
          type: string
        chromosome:
          type: string
        start:
          type: integer
        end:
          type: integer
        description:
          type: string
        expressions:
          type: array
          items:
            $ref: "#/components/schemas/GeneExpressionEntry"
        tissues:
          type: array
          items:
            type: string
        maxTpm:
          type: number
        medianExpressionHistogram:
          type: array
          items:
            type: object
            properties:
              bucket:
                type: string
              count:
                type: integer
            required: [bucket, count]
        gtexUrl:
          type: string
        ensemblUrl:
          type: string
      required: [gene, expressions, tissues]
...
    SearchResultItem:
      type: object
      properties:
        id:
          type: string
        collection:
          type: string
        rowId:
          type: string
        score:
          type: number
        snippet:
          type: string
        fields:
          type: object
          additionalProperties: true
        links:
          type: object
          additionalProperties:
            type: string
      required: [id, collection, rowId, score, fields]
    SearchResponse:
      type: object
      properties:
        query:
          type: string
        total:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResultItem"
      required: [query, total, results]
...
    SearchFacetsResponse:
      type: object
      properties:
        field:
          type: string
        facets:
          type: array
          items:
            type: object
            properties:
              key:
                type: string
              count:
                type: integer
            required: [key, count]
        total:
          type: integer
      required: [field, facets, total]
...
    Citation:
      type: object
      properties:
        title:
          type: string
        journal:
          type: string
        year:
          type: string
        pmid:
          type: string
        doi:
          type: string
        pubmedUrl:
          type: string
        doiUrl:
          type: string
...